improve move functions in lib folder
