+++ ex00/Makefile
@@ -7,7 +7,7 @@
 
 SHELL = /bin/sh
 
-SRC = main.c ourheader.c
+SRC = main.c lib/ft_putstr.c lib/ft_strlen.c lib/ft_is_number.c lib/ft_strncmp.c lib/ft_is_valid_number.c lib/ft_filesize.c lib/ft_file_to_char.c lib/ft_write_result.c
 NAME = rush-02
 
 all: $(NAME)
@@ -17,7 +17,5 @@ $(NAME):
 
 clean:
 
-
 fclean: clean
 	rm -f $(NAME)
-
