+++ ex00/lib/ft_putstr.c
@@ -1,0 +1,19 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_putstr.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: adaferna <adaferna@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2026/02/15 15:03:00 by adaferna          #+#    #+#             */
+/*   Updated: 2026/02/15 15:03:01 by adaferna         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "ourheader.h"
+
+void	ft_putstr(char *str)
+{
+	while(*str)
+		write(1, str++, 1);
+}
