+++ ex00/lib/ft_strlen.c
@@ -1,0 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strlen.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: adaferna <adaferna@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2026/02/15 15:02:54 by adaferna          #+#    #+#             */
+/*   Updated: 2026/02/15 15:03:09 by adaferna         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "ourheader.h"
+
+int	ft_strlen(char *str)
+{
+	int i = 0;
+
+	while(str[i])
+		i++;
+	return(i);
+}
